ta-bot_backend/app/requirements.txt

Flask==3.0.0
Flask-Cors==4.0.0
Flask-Session==0.5.0
openai==0.28.1
pymongo==4.5.0
python-dotenv==1.0.0
redis==5.0.1
gunicorn

